COG-332: UsersSearchCriteria is renamed to CognitoUsersSearchCriteria and moved to cognito package
